+++ src/github/web.rs
@@ -12,6 +12,7 @@ use chrono::{DateTime, Datelike, Days, FixedOffset, TimeZone, Utc};
 use duckdb::arrow::{datatypes::Schema, ipc::writer::FileWriter};
 use minijinja::context;
 use serde::Deserialize;
+use tracing::error;
 
 use crate::{
     web::{render, AppState},
@@ -43,13 +44,21 @@ pub fn data_routes() -> Router<Arc<AppState>, Body> {
 }
 
 async fn fetch_source(State(state): State<Arc<AppState>>) -> AppResult<Html<String>> {
-    let timestamp = fetch_all(&state.pool).await?;
+    let result = fetch_all(&state.pool).await;
 
+    let message = match result {
+        Ok(timestamp) => timestamp.format("%Y-%m-%dT%H:%M:%SZ").to_string(),
+        Err(err) => {
+            let msg = format!("{err}");
+            error!(msg);
+            msg
+        },
+    };
     Ok(Html(render(
         state,
         "sources/fetch_source.html",
         context! {
-            timestamp => timestamp.format("%Y-%m-%dT%H:%M:%SZ").to_string(),
+            message,
         },
     )?))
 }